requirements.txt
amqp==1.4.9
anyjson==0.3.3
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
bcrypt==3.1.4
beautifulsoup4==4.7.1
billiard==3.3.0.23
bs4==0.0.1
celery==3.1.26.post2
celery-with-redis==3.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
colorama==0.4.0
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
cycler==0.10.0
defusedxml==0.5.0
demjson==2.2.4
Django==2.0.6
django-allauth==0.38.0
django-appconf==1.0.2
django-celery==3.2.2
django-celery-results==1.0.4
django-debug-toolbar==1.10.1
django-haystack==2.8.1
django-imagekit==4.0.2
django-mdeditor==0.1.14
django-redis==4.9.0
django-suit==0.2.26
emoji==0.5.1
Fabric3==1.14.post1
gunicorn==19.9.0
httplib2==0.11.3
hyperlink==18.0.0
idna==2.7
image==1.5.27
imagekit==1.0.6
incremental==17.5.0
jieba==0.39
Jinja2==2.10.1
kiwisolver==1.0.1
kombu==3.0.37
lxml==4.3.3
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mysqlclient==1.4.2.post1
npm==0.1.1
numpy==1.16.3
oauthlib==3.0.1
optional-django==0.1.0
paramiko==2.4.2
pilkit==2.0
Pillow==6.0.0
prettytable==0.7.2
pyasn1==0.4.5
pycparser==2.19
pyecharts==1.0.0
PyNaCl==1.3.0
pyparsing==2.4.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
redis==2.10.6
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
urllib3==1.24.2
vine==1.3.0
